+++ plugins/de.cognicrypt.core/src/de/cognicrypt/client/DeveloperProject.java
@@ -8,7 +8,7 @@
  * SPDX-License-Identifier: EPL-2.0
  ********************************************************************************/
 
-package de.cognicrypt.codegenerator;
+package de.cognicrypt.client;
 
 import java.lang.reflect.Method;
 import java.util.Arrays;
@@ -24,7 +24,7 @@ import org.eclipse.jdt.core.IJavaProject;
 import org.eclipse.jdt.core.IPackageFragment;
 import org.eclipse.jdt.core.IPackageFragmentRoot;
 import org.eclipse.jdt.core.JavaCore;
-
+import de.cognicrypt.core.Activator;
 import de.cognicrypt.core.Constants;
 
 /**
@@ -57,17 +57,15 @@ public class DeveloperProject {
 		if (this.project.isOpen() && this.project.hasNature(Constants.JavaNatureID)) {
 			final IJavaProject projectAsJavaProject = JavaCore.create(this.project);
 			final LinkedHashSet<IClasspathEntry> classPathEntryList = new LinkedHashSet<>();
-
 			classPathEntryList.addAll(Arrays.asList(projectAsJavaProject.getRawClasspath()));
 			classPathEntryList.add(JavaCore.newLibraryEntry(this.project.getFile(pathToJar).getFullPath(), null, null, false));
-
 			projectAsJavaProject.setRawClasspath(classPathEntryList.toArray(new IClasspathEntry[1]), null);
 			return true;
 		}
 
 		return false;
 	}
-
+		
 	/**
 	 * Retrieves folder from developer package
 	 * 
@@ -97,7 +95,7 @@ public class DeveloperProject {
 	 *        name of the package
 	 * @return package as {@link org.eclipse.jdt.core.IPackageFragment IPackageFragment}
 	 * @throws CoreException
-	 *         see {@link de.cognicrypt.codegenerator.DeveloperProject#getSourcePath() getSourcePath()}
+	 *         see {@link de.cognicrypt.client.DeveloperProject#getSourcePath() getSourcePath()}
 	 */
 	public IPackageFragment getPackagesOfProject(final String name) throws CoreException {
 		return JavaCore.create(this.project).getPackageFragmentRoot(this.project.getFolder(getSourcePath())).getPackageFragment(name);
